Add a progressbar feature and the corresponding APIs
We now support a new set of APIs :
- set_progress () will create a progressbar, which status
will be determined by a float value (0.0=0%, 1.0=100%) ;
- get_progress () will retrieve the current float value ;
- set_progress_callback () takes a callback function and
a non-mandatory user-supplied pointer as arguments. This
function will be called as often as possible, and is
supposed to modify the progressbar state after blocking
wlmessage_show() has been issued, and call
wlmessage_destroy () when it is finished.

If a callback is given, the textfield and buttons will not
be shown, as the window is then supposed non-interactive.

X11 does not support these APIs for now ; they will have no
effect. Next commit will fix that.

Change-Id: I2ac4cab9b4bb0ed63fdd15baa50ec04b1c8ca769
Signed-off-by: Manuel Bachmann <manuel.bachmann@open.eurogiciel.org>